Rename pass to module
